requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.0
supabase==2.0.0
requests==2.31.0
python-dateutil==2.8.2
twilio==8.10.0
sendgrid==6.11.0
stripe==8.0.0
gunicorn==21.2.0
werkzeug==3.0.1
psycopg2-binary==2.9.9
reportlab==4.0.7
pypdf==4.0.0
apscheduler==3.10.4
